Complete refactor redux login logout with auth0
