+++ public/index.html
@@ -18,6 +18,5 @@
     <div id="react-app"></div>
       <script src="https://cdn.jsdelivr.net/react/0.14.0-rc1/react.js"></script>
       <script src="https://cdn.jsdelivr.net/react/0.14.0-rc1/react-dom.js"></script>
-      <script src="https://cdn.auth0.com/js/lock/11.x/lock.min.js"></script>
   </body>
 </html>
