+++ src/components/Callback.js
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { setIdToken, setAccessToken } from './../utils/Authservice';
+import { setIdToken, setAccessToken } from '../actions/action';
 
 class Callback extends Component {
